\name{nlgm-package}
\alias{nlgm-package}
\docType{package}
\title{
This is an R package that fits 6 non linear growth models.
}
\description{
Six growth models are fitted using non-linear least squares. The relevant reference is:
Reddy T., Shkedy Z., van Rensburg C. J., Mwambi H., Debba P., Zuma K. and Manda, S. (2021). Short-term real-time prediction of total number of reported COVID-19 cases and deaths in South Africa: a data driven approach. BMC Medical Research methodology, 21(1): 1--11. <doi:10.1186/s12874-020-01165-x>.
}
\details{
\tabular{lll}{
Package: \tab nlgmcr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2021-12-01\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr} and Nikolaos Pandis \email{npandis@yahoo.com}.
}

\section{Maintainers }{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Reddy T., Shkedy Z., van Rensburg C. J., Mwambi H., Debba P., Zuma K. and Manda, S. (2021).
Short-term real-time prediction of total number of reported COVID-19 cases and deaths in South Africa: a data driven approach. BMC medical Research Methodology, 21(1): 1--11.
}

%\note{
%\bold{Acknowledgments:}
%}

\keyword{ growth models }



